Added connection to hosted database + fixed variable type inconsistency

+++ src/main/java/gruppe_12_backend/rest_api_12/repository/StrainRepo.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface StrainRepo extends CrudRepository<Strain,Integer> {
-    Strain findById(int id);
+    Strain findById(long id);
 }
